Migare docs (#3512)
