+++ packages/terra-core-docs/src/terra-dev-site/doc/dropdown-button/DropdownButton.1.doc.mdx
@@ -1,10 +1,10 @@
 import { Badge } from 'terra-dropdown-button/package.json?dev-site-package';
 
-import DefaultDropdownButton from '../example/DefaultDropdownButton?dev-site-example';
-import EmphasisDropdownButton from '../example/EmphasisDropdownButton?dev-site-example';
-import GhostDropdownButton from '../example/GhostDropdownButton?dev-site-example';
-import DisabledDropdownButton from '../example/DisabledDropdownButton?dev-site-example';
-import BlockDropdownButton from '../example/BlockDropdownButton?dev-site-example';
+import DefaultDropdownButton from './example/DefaultDropdownButton?dev-site-example';
+import EmphasisDropdownButton from './example/EmphasisDropdownButton?dev-site-example';
+import GhostDropdownButton from './example/GhostDropdownButton?dev-site-example';
+import DisabledDropdownButton from './example/DisabledDropdownButton?dev-site-example';
+import BlockDropdownButton from './example/BlockDropdownButton?dev-site-example';
 
 
 import DropdownButtonPropsTable from 'terra-dropdown-button/src/DropdownButton?dev-site-props-table';
