+++ packages/terra-core-docs/src/terra-dev-site/doc/dropdown-button/SplitButton.2.doc.mdx
@@ -1,9 +1,9 @@
 import { Badge } from 'terra-dropdown-button/package.json?dev-site-package';
 
-import DefaultSplitButton from '../example/DefaultSplitButton?dev-site-example';
-import GhostSplitButton from '../example/GhostSplitButton?dev-site-example';
-import DisabledSplitButton from '../example/DisabledSplitButton?dev-site-example';
-import BlockSplitButton from '../example/BlockSplitButton?dev-site-example';
+import DefaultSplitButton from './example/DefaultSplitButton?dev-site-example';
+import GhostSplitButton from './example/GhostSplitButton?dev-site-example';
+import DisabledSplitButton from './example/DisabledSplitButton?dev-site-example';
+import BlockSplitButton from './example/BlockSplitButton?dev-site-example';
 
 import SplitButtonPropsTable from 'terra-dropdown-button/src/SplitButton?dev-site-props-table';
 import ItemPropsTable from 'terra-dropdown-button/src/Item?dev-site-props-table';
