+++ packages/terra-core-docs/src/terra-dev-site/doc/dropdown-button/example/GhostSplitButton.jsx
