+++ packages/terra-core-docs/src/terra-dev-site/test/dropdown-button/CallbackSplitButton.test.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import classnames from 'classnames/bind';
-import DropdownButton, { Item } from '../../../DropdownButton';
+import { SplitButton, Item } from 'terra-dropdown-button';
 import styles from './CallbackButtons.module.scss';
 
 const cx = classnames.bind(styles);
@@ -10,11 +10,11 @@ const CallbackDropdownButtons = () => {
 
   return (
     <div className={cx('background', color)}>
-      <DropdownButton label="Colors">
+      <SplitButton primaryOptionLabel="Gray" onSelect={() => { setColor('gray'); }}>
         <Item label="Red" onSelect={() => { setColor('red'); }} id="red" />
         <Item label="White" onSelect={() => { setColor('white'); }} id="white" />
         <Item label="Blue" onSelect={() => { setColor('blue'); }} id="blue" />
-      </DropdownButton>
+      </SplitButton>
     </div>
   );
 };
