+++ packages/terra-core-docs/src/terra-dev-site/test/dropdown-button/DefaultDropdownButton.test.jsx
@@ -1,20 +1,20 @@
 import React, { useState } from 'react';
 import classnames from 'classnames/bind';
-import SplitButton, { Item } from '../../../SplitButton';
+import DropdownButton, { Item } from 'terra-dropdown-button';
 import styles from './ExtraSpacing.module.scss';
 
 const cx = classnames.bind(styles);
 
-const DefaultSplitButton = () => {
+const DefaultDropdownButton = () => {
   const [message, setMessage] = useState(' No option clicked');
 
   return (
     <div className={cx('container-spacing-wrapper')}>
-      <SplitButton primaryOptionLabel="Split" metaData={{ key: 'primary-button' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} id="split">
+      <DropdownButton label="Dropdown" id="dropdown">
         <Item id="opt1" label="1st" metaData={{ key: '1st Option' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} />
         <Item id="opt2" label="2nd" metaData={{ key: '2nd Option' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} />
         <Item id="opt3" label="3rd" metaData={{ key: '3rd Option' }} onSelect={(event, metaData) => { setMessage(` ${metaData.key}`); }} />
-      </SplitButton>
+      </DropdownButton>
       <p>
         MetaData of :
         {message}
@@ -23,5 +23,4 @@ const DefaultSplitButton = () => {
   );
 };
 
-export default DefaultSplitButton;
-
+export default DefaultDropdownButton;
