+++ packages/terra-core-docs/src/terra-dev-site/test/dropdown-button/SplitFocus.test.jsx
@@ -1,6 +1,6 @@
 import React, { useRef } from 'react';
 import classnames from 'classnames/bind';
-import SplitButton, { Item } from '../../../SplitButton';
+import { SplitButton, Item } from 'terra-dropdown-button';
 import styles from './ExtraSpacing.module.scss';
 
 const cx = classnames.bind(styles);
