+++ packages/terra-core-docs/src/terra-dev-site/test/dropdown-button/VariantDropdownButtons.test.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 import React from 'react';
 import classnames from 'classnames/bind';
-import DropdownButton, { Item } from '../../../DropdownButton';
+import DropdownButton, { Item } from 'terra-dropdown-button';
 import styles from './ExtraSpacing.module.scss';
 
 const cx = classnames.bind(styles);
