+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Migrate doc to terra-core-docs.
+
 ## 1.29.0 - (September 21, 2021)
 
 * Changed
